upgrade: splash view
