+++ client/lib/core/views/splash_view/splash.view.dart
@@ -1,44 +1,16 @@
-import 'dart:async';
-
-import 'package:client/core/views/signin_view/signin_account_view.dart';
+import 'package:client/core/views/splash_view/splash.viewmodel.dart';
 import 'package:client/core/views/splash_view/splash.widgets.dart';
-import 'package:client/core/views/splash_view/widgets/splash.logo.dart';
 import 'package:flutter/material.dart';
+import 'package:get_it/get_it.dart';
 
-class SplashView extends StatefulWidget with SplashWidgets {
+class SplashView extends StatelessWidget with SplashWidgets {
   SplashView({
     Key? key,
   }) : super(key: key);
-  @override
-  State<SplashView> createState() => _SplashViewState();
-}
+  final _vm = GetIt.I.get<SplashViewModel>();
 
-class _SplashViewState extends State<SplashView> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: FutureBuilder(
-        future: Future.delayed(const Duration(seconds: 4)),
-        builder: (ctx, timer) => timer.connectionState == ConnectionState.done
-            ? const SignInAccountView()
-            : Column(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Spacer(
-                    flex: 3,
-                  ),
-                  const Expanded(
-                    flex: 2,
-                    child: SplashLogo(),
-                  ),
-                  const Spacer(
-                    flex: 2,
-                  ),
-                  Expanded(child: widget.allRightReservedText(context)),
-                  const Spacer()
-                ],
-              ),
-      ),
-    );
+    return _vm.baseScaffold(context, body: body(context));
   }
 }
