+++ client/lib/core/views/splash_view/splash.viewmodel.dart
@@ -1,8 +1,12 @@
+import 'package:client/core/views/abstractions/base_view_model.dart';
 import 'package:injectable/injectable.dart';
 
 @Injectable()
-class SplashViewModel {
+class SplashViewModel extends BaseViewModel {
   splashDelay() {
     Future.delayed(const Duration(milliseconds: 3000), () {});
   }
+
+  @override
+  void clear() {}
 }
