+++ client/lib/core/views/splash_view/splash.widgets.dart
@@ -1,7 +1,9 @@
 import 'package:client/core/utils/constants/colors/color_constans.dart';
 import 'package:client/core/l10n/app_l10n.dart';
 import 'package:client/core/views/common/widgets/text/custom_text.dart';
+import 'package:client/core/views/splash_view/widgets/splash.logo.dart';
 import 'package:flutter/material.dart';
+import '../signin_view/signin_account_view.dart';
 
 class SplashWidgets {
   Widget allRightReservedText(BuildContext context) {
@@ -15,6 +17,31 @@ class SplashWidgets {
           CustomText('2019', color: ColorConstant.instance.grey),
         ],
       ),
+    );
+  }
+
+  Widget body(BuildContext context) {
+    return FutureBuilder(
+      future: Future.delayed(const Duration(seconds: 4)),
+      builder: (ctx, timer) => timer.connectionState == ConnectionState.done
+          ? const SignInAccountView()
+          : Column(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                const Spacer(
+                  flex: 3,
+                ),
+                const Expanded(
+                  flex: 2,
+                  child: SplashLogo(),
+                ),
+                const Spacer(
+                  flex: 2,
+                ),
+                Expanded(child: allRightReservedText(context)),
+                const Spacer()
+              ],
+            ),
     );
   }
 }
